requirements.txt
evernote==1.25.1
httplib2==0.9.2
oauth2==1.9.0.post1
PyYAML==3.11
